+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove follow relationships when receiving `Block` activity.
 - Support `Add(Delete)` activities.
 - Support followers-only reposts.
+- Support WebFinger addresses with IPv6 hostnames.
 
 ### Changed